Move density_to_mat to example_geometry.py
